Chore: Typescript: Rewrite MultiIcon (#21016)
