+++ assets/js/components/MultiIcon/index.ts
